redis should report as a loaded extension
Redis has been implemented in its entirety in PHP by @sgolemon,
and there currently is no way of specifying that it should show
as a loaded extension.

This is a temporary workaround (such as the APC one) to just return true
for extension_loaded('redis') when the Redis class exists. Eventually
this should probably be done from PHP, such as done with
IMPLEMENT_DEFAULT_EXTENSION in C++.